+++ types/object/getAtJsonPointer.d.ts
@@ -1,0 +1,1 @@
+export { getAtJsonPointer, getAtJsonPointer as default } from "../object.js";