+++ types/object/hasAtJsonPointer.d.ts
@@ -1,0 +1,1 @@
+export { hasAtJsonPointer, hasAtJsonPointer as default } from "../object.js";