+++ types/object/parseJsonPointer.d.ts
@@ -1,0 +1,1 @@
+export { parseJsonPointer, parseJsonPointer as default } from "../object.js";